requirements.txt
fredapi
pandas
numpy
matplotlib
seaborn
statsmodels
python-dotenv